Tightened up both functions
Got rid of some unnecessary directory changes, added the user_input as an argument for the input_logic function, which is then plugged in as the path arguments in both resize_photo and make_folder. This provided great opportunity to clean up the functions. Added some more line breaks for aesthetic purposes.