Logic cleanup
Removed double import of the PIL module, added extra logic for reducing user input error, used normpath on user_input to convert the input to be usable on any system, cleaned up some variable naming conventions, moved the help menu to its own function for display. Overall cleanup of the functions that are present.

Additionally changed the readme to match the new functionality.