Changed name, logic and exception refactoring
Changed the name of the program to match the use-case. Got rid of some unnecessary exception handling, moved hard-coded variables to be defined at the top of the program, added a lot of user feedback/prompts. Changed how the instructions are laid out to provide a clearer idea of expected behavior and what the tool is for.